Added Static File Server for Viewing on Local Network
The staticServer allows files in the /var/lib/infinitive folder to be viewed on the local network.
Source code based on examples found on-lone with minimal adaptation.
staticServer should be installed with systemd in /var/lib/staticServer. 
dayLinks.html is now index.html and in place of file path links, the links are used by staticServer.
Minor fixes in infinitive as well as matching code changes to use the staticServer.